Initial version of an Exchange for SSL
This is a first cut at a version of an SSL Exchange. My concerns
around this implementation are in the possibility of wrap being
called at the same time by the send and receive sides of the process.
The javadocs mention that wrap or unwrap should only be called from
one thread.

Also, this version does not properly clean up the engine when the
read or write are closed.